Implementing asyncio: 1. Added aiosqlite as a dependency
This is a necessary dependency to use asyncio with the sqlite3 module
because sqlite3 does not implement awaitable.

2. Creating a function "perform_db_operations"

This function is responsible for running all of the database methods
(such as connecting, getting the cursor, and reading from the
database). if the user calls determine-relatedness then it just
basically uses a single thread to determine the cases. If they call
gather-distributions then it will use asyncio to gather the controls and
cases.

3. Created a dbResults class

This class has the database path, the database table name, and
attributes for the cases relatedness and the controls relatedness from
the database. The final two attributes will be list of tuple with
information about the ID1, ID2, and the estimated relatedness. If the
user calls the determine-relatedness command then only the case_results
attribute gets updated and the control_results attribute will be None.

4 Updated the write to file method

Moved the write_to_file method to a file called writer.py in the utility
module. It also now accepts the dbResults object instead of a results
list.