Added function to find all connections or only grids in file: --- 1. Added the flag --all-connections:
This flag indicates that the sql query should be a OR not and AND. A OR
query will return all pairs where at least 1 individual is in the grid
file.

2. Added a scalene dev dependency:

3. Abstracted away the file handling for the grid file:

Created a class called FileReader that will open the grid file and then
read in the grids. It has methods __enter__ and __exit__ so that the
resource can be closed with a context manager. It will raise a OSError
if the file can't be opened. It will also raise a
utilities.IncorrectGridFileFormat if the file is not properly formatted